Restore generated type registration test coverage
