+++ example/enums/query.sql_test.go
@@ -16,6 +16,7 @@ func TestNewQuerier_FindAllDevices(t *testing.T) {
 	defer cancel()
 	conn, cleanup := pgtest.NewPostgresSchema(t, []string{"schema.sql"})
 	defer cleanup()
+	require.NoError(t, RegisterTypes(ctx, conn))
 
 	q := NewQuerier(conn)
 	mac, _ := net.ParseMAC("00:00:5e:00:53:01")
@@ -49,6 +50,7 @@ func TestNewQuerier_FindOneDeviceArray(t *testing.T) {
 	defer cancel()
 	conn, cleanup := pgtest.NewPostgresSchema(t, []string{"schema.sql"})
 	defer cleanup()
+	require.NoError(t, RegisterTypes(ctx, conn))
 
 	q := NewQuerier(conn)
 
@@ -64,6 +66,7 @@ func TestNewQuerier_FindManyDeviceArray(t *testing.T) {
 	defer cancel()
 	conn, cleanup := pgtest.NewPostgresSchema(t, []string{"schema.sql"})
 	defer cleanup()
+	require.NoError(t, RegisterTypes(ctx, conn))
 
 	q := NewQuerier(conn)
 
@@ -79,6 +82,7 @@ func TestNewQuerier_FindManyDeviceArrayWithNum(t *testing.T) {
 	defer cancel()
 	conn, cleanup := pgtest.NewPostgresSchema(t, []string{"schema.sql"})
 	defer cleanup()
+	require.NoError(t, RegisterTypes(ctx, conn))
 
 	q := NewQuerier(conn)
 	one, two := int32(1), int32(2)
@@ -98,6 +102,7 @@ func TestNewQuerier_EnumInsideComposite(t *testing.T) {
 	defer cancel()
 	conn, cleanup := pgtest.NewPostgresSchema(t, []string{"schema.sql"})
 	defer cleanup()
+	require.NoError(t, RegisterTypes(ctx, conn))
 
 	q := NewQuerier(conn)
 	mac, _ := net.ParseMAC("08:00:2b:01:02:03")
