+++ example/function/query.sql_test.go
@@ -13,6 +13,7 @@ import (
 func TestNewQuerier_OutParams(t *testing.T) {
 	conn, cleanup := pgtest.NewPostgresSchema(t, []string{"schema.sql"})
 	defer cleanup()
+	require.NoError(t, RegisterTypes(t.Context(), conn))
 
 	q := NewQuerier(conn)
